Added basic styling to register form
