+++ package.json
@@ -27,6 +27,7 @@
     "@testing-library/jest-dom": "5.16.5",
     "@testing-library/react": "13.4.0",
     "@testing-library/user-event": "14.4.3",
+    "@types/bcrypt": "^5.0.0",
     "@types/jest": "29.2.1",
     "@types/node": "18.11.6",
     "@types/react": "18.0.24",
