+++ src/pages/account/register/Register.module.scss
@@ -1,0 +1,27 @@
+.container {
+  display: flex;
+  flex-direction: column;
+}
+
+.registerForm {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  border: 3px solid black;
+  padding: 2.5em;
+  border-radius: 15px;
+
+  & input {
+    width: 100%;
+    height: 3.25em;
+    border-radius: 10px;
+    margin: 1em 0;
+    padding-left: 1.5em;
+  }
+
+  & button {
+    height: 4.5em;
+    border-radius: 10px;
+    margin: 2.5em 0;
+  }
+}
